second order  : params dipole

+++ experiment_fig-secondorder_segmentation.py
@@ -1,7 +1,7 @@
 
 # rm **/Geisler01Fig7A*
 
-
+import os
 import __init__
 import numpy as np
 from NeuroTools.parameters import ParameterSet
@@ -18,7 +18,8 @@ pe.n_theta = 48
 pe.do_whitening = True
 
 
-figpath = '../../CNRS/BICV-book/BICV_INT/BICV-sparse/'
+#figpath = '../../CNRS/BICV-book/BICV_INT/BICV-sparse/'
+figpath = './'
 FORMATS = ['pdf', 'eps']
 fig_width_pt = 318.670  # Get this from LaTeX using \showthe\columnwidth
 inches_per_pt = 1.0/72.27               # Convert pt to inches
@@ -46,42 +47,54 @@ print mp.n_levels, mp.sf_0
 
 print ' without second-order '
 matname = 'mat/Geisler01Fig7A.npy'
+if not(os.path.isfile(matname)):
+    if not(os.path.isfile(matname + '_lock')):
+        file(matname + '_lock', 'w').close()
+        mp.pe.eta_SO = 0.
+        edges, C_res = mp.run_mp(image, verbose=True)
+        np.save(matname, edges)
+        os.remove(matname + '_lock')
 try:
     edges = np.load(matname)
+    fig, a = mp.show_edges(edges, image=image, v_min=v_min, v_max=v_max, color='toto', show_phase=False) #
+    if not(figpath==None): 
+        for ext in FORMATS: fig.savefig(figpath + 'Geisler01Fig7_A.' + ext)
 except:
-    mp.pe.eta_SO = 0.
-    edges, C_res = mp.run_mp(image, verbose=True)
-    np.save(matname, edges)    
-
-edges = np.load(matname)
-fig, a = mp.show_edges(edges, image=image, v_min=v_min, v_max=v_max, color='toto', show_phase=False) #
-if not(figpath==None): 
-    for ext in FORMATS: fig.savefig(figpath + 'Geisler01Fig7_A.' + ext)
-        
+    print matname
+    
 print ' with second-order '
 matname = 'mat/Geisler01Fig7A_secondorder.npy'
+if not(os.path.isfile(matname)):
+    if not(os.path.isfile(matname + '_lock')):
+        file(matname + '_lock', 'w').close()
+        mp.pe.eta_SO = 0.75
+        edges, C_res = mp.run_mp(image, verbose=True)
+        np.save(matname, edges)
+        os.remove(matname + '_lock')
 try:
     edges = np.load(matname)
+    edges[4, :] *= -1 # turn red in blue...
+    fig, a = mp.show_edges(edges, image=image, v_min=v_min, v_max=v_max, color='toto', show_phase=False) #
+    if not(figpath==None): 
+        for ext in FORMATS: fig.savefig(figpath + 'Geisler01Fig7_B.' + ext)
 except:
-    mp.pe.eta_SO = .75
-    edges, C_res = mp.run_mp(image, verbose=True)
-    np.save(matname, edges)    
+    print matname
 
-edges = np.load(matname)
-edges[4, :] *= -1 # turn red in blue...
-fig, a = mp.show_edges(edges, image=image, v_min=v_min, v_max=v_max, color='toto', show_phase=False) #
-if not(figpath==None): 
-    for ext in FORMATS: fig.savefig(figpath + 'Geisler01Fig7_B.' + ext)
 
 if True:    
     for mp.pe.eta_SO in np.linspace(.05, 1., 9):
         matname = 'mat/Geisler01Fig7A_secondorder_' + str(mp.pe.eta_SO).replace('.', '_') + '.npy'
+        if not(os.path.isfile(matname)):
+            if not(os.path.isfile(matname + '_lock')):
+                file(matname + '_lock', 'w').close()
+                edges, C_res = mp.run_mp(image, verbose=True)
+                np.save(matname, edges)
+                os.remove(matname + '_lock')
         try:
             edges = np.load(matname)
+            edges[4, :] *= -1 # turn red in blue...
+            fig, a = mp.show_edges(edges, image=image, v_min=v_min, v_max=v_max, color='toto', show_phase=False) #
+            fig.savefig(matname.replace('mat/', mp.pe.figpath).replace('.npy', '.pdf'))
         except:
-            edges, C_res = mp.run_mp(image, verbose=True)
-            np.save(matname, edges)    
-        
-        fig, a = mp.show_edges(edges, image=image, v_min=v_min, v_max=v_max, color='toto', show_phase=False) #
-        fig.savefig(matname.replace('mat/', mp.pe.figpath).replace('.npy', '.pdf'))
+            print matname
